requirements.txt
fastapi
pymongo
dotenv
neo4j
qdrant_client
groq
neo4j_graphrag
requests
neo4j_graphrag[ollama]
langchain_text_splitters
python-multipart
uvicorn
python-docx
pypdf
Pillow
pytesseract

langchain
langchain-openai
langchain-groq
langchain-ollama
langgraph
pydantic
python-dotenv
slowapi
PyJWT
